% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofix.R
\name{add_to_description}
\alias{add_to_description}
\title{Add Package to DESCRIPTION}
\usage{
add_to_description(package, field = "Imports", path = ".")
}
\arguments{
\item{package}{Character. Package name.}

\item{field}{Character. DESCRIPTION field. Default: "Imports".}

\item{path}{Character. Path to project root.}
}
\value{
Logical indicating success
}
\description{
Adds a package to the Imports field in DESCRIPTION.
}
\keyword{internal}
