% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{BASE_PACKAGES}
\alias{BASE_PACKAGES}
\title{Base R Packages}
\usage{
BASE_PACKAGES
}
\description{
Base R packages that don't need explicit declaration in DESCRIPTION.
These packages are always available in R installations.
}
\keyword{internal}
