% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_library_calls}
\alias{extract_library_calls}
\title{Extract library() Calls}
\usage{
extract_library_calls(lines)
}
\arguments{
\item{lines}{Character vector of code lines.}
}
\value{
Character vector of package names
}
\description{
Extract library() Calls
}
\keyword{internal}
