% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_namespace_calls}
\alias{extract_namespace_calls}
\title{Extract Namespace Calls}
\usage{
extract_namespace_calls(lines)
}
\arguments{
\item{lines}{Character vector of code lines.}
}
\value{
Character vector of package names
}
\description{
Extracts package names from pkg::function() syntax.
}
\keyword{internal}
