% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_packages_from_file}
\alias{extract_packages_from_file}
\title{Extract Packages from Single File}
\usage{
extract_packages_from_file(file, skip_comments = TRUE)
}
\arguments{
\item{file}{Character. Path to R file.}

\item{skip_comments}{Logical. Skip commented lines.}
}
\value{
Character vector of package names
}
\description{
Extracts package references from a single R file.
}
\keyword{internal}
