% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_require_calls}
\alias{extract_require_calls}
\title{Extract require() Calls}
\usage{
extract_require_calls(lines)
}
\arguments{
\item{lines}{Character vector of code lines.}
}
\value{
Character vector of package names
}
\description{
Extract require() Calls
}
\keyword{internal}
