% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_roxygen_imports}
\alias{extract_roxygen_imports}
\title{Extract Roxygen Imports}
\usage{
extract_roxygen_imports(lines)
}
\arguments{
\item{lines}{Character vector of code lines.}
}
\value{
Character vector of package names
}
\description{
Extracts package names from roxygen2 @importFrom and @import tags.
}
\keyword{internal}
