% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{FILE_EXTENSIONS}
\alias{FILE_EXTENSIONS}
\title{File Extensions to Search}
\format{
An object of class \code{character} of length 4.
}
\usage{
FILE_EXTENSIONS
}
\description{
R source file extensions to scan for package references.
}
\keyword{internal}
