% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{filter_skip_paths}
\alias{filter_skip_paths}
\title{Filter Skip Paths}
\usage{
filter_skip_paths(files)
}
\arguments{
\item{files}{Character vector of file paths.}
}
\value{
Filtered character vector
}
\description{
Removes files matching skip patterns.
}
\keyword{internal}
