% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{find_r_files}
\alias{find_r_files}
\title{Find R Files in Directories}
\usage{
find_r_files(dirs, path = ".")
}
\arguments{
\item{dirs}{Character vector of directory names.}

\item{path}{Character. Base path.}
}
\value{
Character vector of file paths
}
\description{
Recursively finds R source files in specified directories.
}
\keyword{internal}
