% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{format_r_vector}
\alias{format_r_vector}
\title{Format R Vector}
\usage{
format_r_vector(packages)
}
\arguments{
\item{packages}{Character vector.}
}
\value{
Character string like 'c("pkg1", "pkg2")'
}
\description{
Formats package names as R vector string.
}
\keyword{internal}
