% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofix.R
\name{handle_auto_fix_description}
\alias{handle_auto_fix_description}
\title{Handle Auto-Fix for DESCRIPTION}
\usage{
handle_auto_fix_description(packages, path = ".")
}
\arguments{
\item{packages}{Character vector of package names to add.}

\item{path}{Character. Path to project root.}
}
\value{
Character vector of packages that failed to add
}
\description{
Automatically adds missing packages to DESCRIPTION Imports.
}
\keyword{internal}
