% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{has_description}
\alias{has_description}
\title{Check if DESCRIPTION Exists}
\usage{
has_description(path = ".")
}
\arguments{
\item{path}{Character. Path to project root.}
}
\value{
Logical.
}
\description{
Check if DESCRIPTION Exists
}
\keyword{internal}
