% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{has_renv_lock}
\alias{has_renv_lock}
\title{Check if renv.lock Exists}
\usage{
has_renv_lock(path = ".")
}
\arguments{
\item{path}{Character. Path to project root.}
}
\value{
Logical.
}
\description{
Check if renv.lock Exists
}
\keyword{internal}
