% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{remove_from_renv_lock}
\alias{remove_from_renv_lock}
\title{Remove Packages from renv.lock}
\usage{
remove_from_renv_lock(packages, path = ".")
}
\arguments{
\item{packages}{Character vector. Package names to remove.}

\item{path}{Character. Path to project root.}
}
\value{
Character vector of successfully removed packages
}
\description{
Removes specified packages from renv.lock.
}
