% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{show_manual_fix_lock_instructions}
\alias{show_manual_fix_lock_instructions}
\title{Show Manual Fix Lock Instructions}
\usage{
show_manual_fix_lock_instructions(packages)
}
\arguments{
\item{packages}{Character vector of missing packages.}
}
\description{
Show Manual Fix Lock Instructions
}
\keyword{internal}
