% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{SKIP_DIRS}
\alias{SKIP_DIRS}
\title{Directories to Skip During Scanning}
\format{
An object of class \code{character} of length 3.
}
\usage{
SKIP_DIRS
}
\description{
Directory patterns to exclude from package scanning.
}
\keyword{internal}
