% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{STRICT_DIRS}
\alias{STRICT_DIRS}
\title{Strict Mode Directories}
\usage{
STRICT_DIRS
}
\description{
All directories scanned in strict mode (includes tests and
vignettes). See \code{STANDARD_DIRS} for why "." is not listed.
}
\keyword{internal}
