% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_bioconductor}
\alias{validate_bioconductor}
\title{Validate Package on Bioconductor}
\usage{
validate_bioconductor(package)
}
\arguments{
\item{package}{Character. Package name.}
}
\value{
Logical. TRUE if package exists on Bioconductor.
}
\description{
Checks if a package exists on Bioconductor.
}
\keyword{internal}
