% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_cran}
\alias{validate_cran}
\title{Validate Package on CRAN}
\usage{
validate_cran(package)
}
\arguments{
\item{package}{Character. Package name.}
}
\value{
Logical. TRUE if package exists on CRAN.
}
\description{
Checks if a package exists on CRAN.
}
\keyword{internal}
