% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{SKIP_FILES}
\alias{SKIP_FILES}
\title{Files to Skip During Scanning}
\format{
An object of class \code{character} of length 3.
}
\usage{
SKIP_FILES
}
\description{
Documentation and example files that should not be scanned
for package dependencies (often contain example code).
}
\keyword{internal}
