% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{STANDARD_DIRS}
\alias{STANDARD_DIRS}
\title{Standard Directories to Scan}
\format{
An object of class \code{character} of length 4.
}
\usage{
STANDARD_DIRS
}
\description{
Default directories scanned in standard (non-strict) mode.
}
\keyword{internal}
