% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{STRICT_DIRS}
\alias{STRICT_DIRS}
\title{Strict Mode Directories}
\format{
An object of class \code{character} of length 7.
}
\usage{
STRICT_DIRS
}
\description{
All directories scanned in strict mode (includes tests and vignettes).
}
\keyword{internal}
